refactor(activerecord): move remaining support-subject tests to support/
ar-db-forks-parity, canonical-schema-bulk-inbound-fk, handler-resolved-adapter,
and pg-template exercise modules that moved to support/; keep each test next to
its subject.
